LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_after_target_appended_mouse-expected.txt

child
done

PASS pointer events from mouse received before/after child attached at pointerdown
PASS pointer events from mouse received before/after child attached at pointerup
FAIL pointer events from mouse received before/after child moved at pointerdown assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-moved),pointerover@child,pointerenter@child,pointerup@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-moved),pointerover@child,pointerenter@parent,pointerenter@child,pointerup@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent"
FAIL pointer events from mouse received before/after child moved at pointerup assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-moved),pointerover@child,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-moved),pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent"
PASS mouse events from mouse received before/after child attached at mousedown
PASS mouse events from mouse received before/after child attached at mouseup
FAIL mouse events from mouse received before/after child moved at mousedown assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,(child-moved),mouseover@child,mouseenter@child,mouseup@child,mousedown@child,mouseup@child,mouseout@child,mouseleave@child,mouseleave@parent" but got "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,(child-moved),mouseover@child,mouseenter@parent,mouseenter@child,mouseup@child,mousedown@child,mouseup@child,mouseout@child,mouseleave@child,mouseleave@parent"
FAIL mouse events from mouse received before/after child moved at mouseup assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,mouseup@child,(child-moved),mouseover@child,mouseenter@child,mousedown@child,mouseup@child,mouseout@child,mouseleave@child,mouseleave@parent" but got "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,mouseup@child,(child-moved),mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,mouseup@child,mouseout@child,mouseleave@child,mouseleave@parent"
